Actually compiles now, and I hate inlining even more
